fix : 존재하지 않는 카테고리 삭제 시 404오류

+++ src/main/java/com/aoo/admin/adapter/out/persistence/CategoryPersistenceAdapter.java
@@ -48,6 +48,7 @@ public class CategoryPersistenceAdapter implements SaveCategoryPort, FindCategor
 
     @Override
     public DeleteCategoryResult delete(Long categoryId) {
+        if (!categoryJpaRepository.existsById(categoryId)) throw new AdminException(AdminErrorCode.CATEGORY_NOT_FOUND);
         categoryJpaRepository.deleteById(categoryId);
         return DeleteCategoryResult.of(categoryId);
     }
